[crucible-llvm] Split Translation tests to separate module.
