+++ crucible-llvm/crucible-llvm.cabal
@@ -105,7 +105,7 @@ test-suite crucible-llvm-tests
   hs-source-dirs: test
   default-language: Haskell2010
   ghc-options: -Wall -Wcompat
-  -- other-modules:
+  other-modules: TestTranslation
   build-depends:
     base,
     bv-sized,
